requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pymongo==4.6.0
motor==3.3.2
redis==5.0.1
pydantic==2.5.0
pydantic-settings==2.1.0
python-dotenv==1.0.0
bcrypt==4.1.1
pyjwt==2.12.1
python-multipart==0.0.6
email-validator==2.1.0
apscheduler==3.10.4
pytz==2024.1
